decompose validate optimizer command
